Fix long number parse

+++ core/src/main/java/com/soffid/iam/addons/report/service/ejb/ClusterManagerBean.java
@@ -1,70 +1,20 @@
 package com.soffid.iam.addons.report.service.ejb;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.net.Inet4Address;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.nio.file.Files;
-import java.sql.Connection;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.util.Collection;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.jar.JarOutputStream;
-import java.util.zip.ZipEntry;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
-import javax.ejb.Local;
 import javax.ejb.SessionContext;
 import javax.ejb.Singleton;
 import javax.ejb.Startup;
 import javax.ejb.Timeout;
 import javax.ejb.Timer;
-import javax.ejb.TimerConfig;
 
 import org.apache.commons.logging.LogFactory;
-import org.hibernate.SessionFactory;
-import org.hibernate.classic.Session;
 
 import com.soffid.iam.ServiceLocator;
-import com.soffid.iam.addons.report.api.ExecutedReport;
-import com.soffid.iam.addons.report.api.ParameterValue;
-import com.soffid.iam.addons.report.api.Report;
-import com.soffid.iam.addons.report.service.ReportSchedulerService;
-import com.soffid.iam.addons.report.service.ReportService;
 import com.soffid.iam.api.Configuration;
-import com.soffid.iam.doc.api.DocumentOutputStream;
-import com.soffid.iam.doc.api.DocumentReference;
-import com.soffid.iam.doc.exception.DocumentBeanException;
-import com.soffid.iam.doc.service.DocumentService;
 import com.soffid.iam.service.ConfigurationService;
-import com.soffid.iam.utils.Security;
-
-import bsh.EvalError;
-import es.caib.seycon.ng.exception.InternalErrorException;
-import net.sf.jasperreports.engine.JRBand;
-import net.sf.jasperreports.engine.JRChild;
-import net.sf.jasperreports.engine.JRException;
-import net.sf.jasperreports.engine.JRParameter;
-import net.sf.jasperreports.engine.JRSubreport;
-import net.sf.jasperreports.engine.JasperExportManager;
-import net.sf.jasperreports.engine.JasperFillManager;
-import net.sf.jasperreports.engine.JasperPrint;
-import net.sf.jasperreports.engine.JasperReport;
-import net.sf.jasperreports.engine.export.JRCsvExporter;
-import net.sf.jasperreports.engine.export.JRXlsExporter;
-import net.sf.jasperreports.engine.export.JRXlsExporterParameter;
-import net.sf.jasperreports.engine.query.JRHibernateQueryExecuterFactory;
-import net.sf.jasperreports.engine.util.JRLoader;
 
 @Singleton(name="ReportClusterBean")
 @Startup
@@ -105,7 +55,7 @@ public class ClusterManagerBean  {
 				String [] split = cfg.getValue().isEmpty() ? new String[0]: cfg.getValue().split(" ");
 				if (split.length != 2 ||
 						split[0].equals(hostName) ||
-						Integer.parseInt(split[1]) < timeout)
+						Long.parseLong(split[1]) < timeout)
 				{
 					cfg.setValue(hostName+" "+now);
 					svc.update(cfg);
